D3.d complete- Ask for reset confirmation

+++ src/main.ts
@@ -44,7 +44,7 @@ class Cache implements Memento<string> {
 
 const zoomAmount = 19;
 const origin = [36.989498, -122.062777];
-let playerLocation = [origin[0], origin[1]];
+let playerLocation = [...origin];
 const savedLocation = localStorage.getItem("playerLocation");
 if (savedLocation) {
   playerLocation = JSON.parse(savedLocation);
@@ -75,8 +75,10 @@ let path: leaflet.latlng = [[...playerLocation]];
 const savedPath = localStorage.getItem("savedPath");
 if (savedPath) {
   path = JSON.parse(savedPath);
+  console.log("saved", path);
 }
 let polyline = leaflet.polyline(path, { color: "red" }).addTo(map);
+
 function playerController(dir: string, lat: number, lon: number) {
   const button = document.querySelector<HTMLDivElement>(dir)!;
   button.addEventListener("click", () => {
@@ -145,6 +147,7 @@ function resetMap() {
   path.push([...playerLocation]);
   localStorage.setItem("savedPath", JSON.stringify(path));
   polyline.setLatLngs(path);
+
   map.panTo(playerLocation);
   playerMarker.setLatLng(playerLocation);
   map.removeLayer(cacheLayer);
@@ -218,15 +221,23 @@ function populateNeighborhood() {
 
 const reset = document.querySelector<HTMLDivElement>("#reset")!;
 reset.addEventListener("click", () => {
-  localStorage.clear();
-  playerCoins = [];
-  status.innerHTML = `You have 0 coin(s)`;
-  playerLocation = [origin[0], origin[1]];
-  resetMap();
-  polyline.removeFrom(map);
-  path = [[...playerLocation]];
-  polyline = leaflet.polyline(path, { color: "red" }).addTo(map);
+  const confirm = prompt(
+    "Are you sure you want to reset? All progress will be lost (yes/no)",
+  );
+  if (confirm == "yes") {
+    playerCoins = [];
+    status.innerHTML = `You have 0 coin(s)`;
+    playerLocation = [...origin];
+    resetMap();
+    if (polyline) {
+      polyline.removeFrom(map);
+    }
+    path = [[...origin]];
+    polyline = leaflet.polyline(path, { color: "red" }).addTo(map);
+    localStorage.clear();
+  }
 });
+
 const sensor = document.querySelector<HTMLDivElement>("#sensor")!;
 sensor.addEventListener("click", () => {
   if (sensor.classList.contains("locating")) {
